Support extending setup steps with plugin module injection JW8-1596
