+++ src/js/plugins/loader.js
@@ -1,17 +1,5 @@
 import Promise, { resolved } from 'polyfills/promise';
-
-function configurePlugin(pluginObj, pluginConfig, api) {
-    const pluginName = pluginObj.name;
-
-    const div = document.createElement('div');
-    div.id = api.id + '_' + pluginName;
-    div.className = 'jw-plugin jw-reset';
-
-    const pluginOptions = Object.assign({}, pluginConfig);
-    const pluginInstance = pluginObj.getNewInstance(api, pluginOptions, div);
-
-    api.addPlugin(pluginName, pluginInstance);
-}
+import { configurePlugin } from 'plugins/plugin';
 
 const PluginLoader = function () {
     this.load = function (api, pluginsModel, pluginsConfig, model) {
